[SECURITY-1245] Fold tests back into normal files
